Safari-611.1.10/LayoutTests/editing/async-clipboard/clipboard-change-data-while-getting-type-expected.txt

This test verifies that if platform pasteboard contents are changed before reading from a ClipboardItem, the new contents of the pasteboard should not be exposed to the page and the result of getType() should be rejected. This test requires WebKitTestRunner.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS items.length is 1
PASS Handled exception: NotAllowedError
PASS successfullyParsed is true

TEST COMPLETE
